cassandra-table-url-replace/replace_old_url/requirements.txt
cassandra-driver == 3.15.1
Urlparse2 == 1.1.1